pull apart output module for readability
